make products controller

+++ src/main/java/com/example/etl1/repository/ProductRepository.java
@@ -2,5 +2,5 @@ package com.example.etl1.repository;
 import com.example.etl1.model.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Long> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
 }
