+++ src/main/java/com/example/etl1/repository/ShippingLaneRepository.java
@@ -3,5 +3,5 @@ package com.example.etl1.repository;
 import com.example.etl1.model.ShippingLane;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ShippingLaneRepository extends CrudRepository<ShippingLane, Long> {
+public interface ShippingLaneRepository extends CrudRepository<ShippingLane, Integer> {
 }
